--- Complexity/Application/ITaskGeneratorSelector.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-using Domain.Entities.TaskGenerators;
-
-namespace Application
-{
-    public interface ITaskGeneratorSelector
-    {
-        TaskGenerator Select(IEnumerable<TaskGenerator> generators);
-    }
-}
