+++ Complexity/Tests/Mocks/TestTaskGeneratorSelector.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Application;
+using Application.Selectors;
 using Domain.Entities.TaskGenerators;
 
 namespace Tests.Mocks
